pyproject.toml
[project]
name = "llm-intent-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "awscli>=1.43.6",
    "boto3>=1.42.0",
    "botocore>=1.41.6",
    "python-dotenv>=1.2.1",
]
